Removed Useless Middleware;
Removed Local middleware that had no purpose -> In summary, It would get the current set locale of  the app and then change the locale of the app to the locale it got. So In other words, it would change the locale of the app to the locale of the app.